requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mwclient==0.10.0
oauthlib==3.1.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
python-telegram-bot==12.4.2
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
wrapt==1.11.2
